+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+### Changed
+  - [Issue #133](https://github.com/nasa/stitchee/issues/133): Add readthedocs documentation build
+### Deprecated
+### Removed
+### Fixed
+
 ## [1.2.1]
 
 ### Added