+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [Issue #133](https://github.com/nasa/stitchee/issues/133): Add readthedocs documentation build
 - [Issue #185](https://github.com/nasa/stitchee/issues/185): Added arguments for temporary file copies and overwriting output file in main stitchee function
+- [Issue #181](https://github.com/nasa/stitchee/issues/181): Add a group delimiter argument
 ### Changed
 ### Deprecated
 ### Removed